Update Website.py
Snatched this one from the clutches of a backed up home folder on another harddrive. This revision really creates a compelling reason to use a template script.

It has a flexible blocky layout and there is a neat javascript at the bottom that randomizes all colors so its easy to debug.